+++ R/globalVariables.R
@@ -2,4 +2,4 @@
 utils::globalVariables(c(".", "p", "where", "x", "y", "variance.ratio", "V1",
                          "estimate", "estimate1", "estimate2", "df", "Method",
                          "Alternative", "r", "W", "CI", "Fit", "Parameter",
-                         "95% CI", "95% CI (b)", "b"))
+                         "95% CI", "95% CI (b)", "b", "Dependent Variable"))